Converted sitemap to api
